New LOGO 🎉 (#96)
* Add logo donut

* Add bouncing animation (first version)

* Replace project name and logo

* Add final logo images

One image (logo-donut-simple) is for the rolling effect
The other one is for the static logo

* Add standalone logo page for testing

* Add logo text image

* Adjust animation

Using relative positioning, through translate3d()

* Add new animation position and text logo

* Add animation into SVG logo

The animation makes the logo hair pops up at the end

* Use new SVG with animation

* Apply Monserrat font to h1 and h2

* New donut logo

This version is centered on the center of the donut, to allow for
a smoother rotation aninmation. Also the colored part  inside is
slightly more offset towards the bottom right

* Embedded minified SVG logo in page

* Add logo "hair" animation to SCSS

This now applies directly to tghe embedded SVG

* Move enetring point for the bouncing logo

Moved from 50% to 100% of x axis. This makes for a smoother animation.

* Add logo track svg

* Add logo track to index

* Add end of logo animation

* Optmize logo track SVG

And remove old (no longer used) logo SVGs

* Move styles into (S)CSS

* Logo SVG clean-up and optimization

* Remove test file

* fix rebasing misses